requirements.txt
flask == 3.0.3
requests == 2.32.3 
httpx == 0.26.0
fhir.resources == 7.1.0
psycopg2 == 2.9.9
pydantic == 2.8.2